requirements.txt
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
distro==1.9.0
fastapi==0.120.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
numpy==2.2.6
openai==2.6.0
opencv-python==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.48.0
tesserocr==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
